requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.0.11
django-bootstrap4==2.3.1
django-debug-toolbar==3.2
django-extensions==3.1.0
importlib-metadata==2.1.1
Pillow==8.1.0
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
zipp==3.4.0
